Solidify existing UTF-8 decoding behaviour on binary-format text[] (#102)
This never did the right thing, even when it was introduced back in brianc/node-postgres@5b3c501d74d0b3d1610dadf776c746a956537925, reading from the global object instead of the client. It’ll usually be undefined and result in UTF-8 decoding, but when it isn’t…

```
> const {Client} = require('pg');
> const client = new Client();
> await client.connect();
> encoding = 'base64';
> (await client.query({text: "SELECT 'foo' x WHERE $1 = 1", values: [1], binary: true})).rows
[ { x: 'foo' } ]
> (await client.query({text: "SELECT ARRAY['foo'] x WHERE $1 = 1", values: [1], binary: true})).rows
[ { x: [ 'Zm9v' ] } ]
```

UTF-8 matches the non-array binary parser for the text type.